requirements.txt
setuptools~=57.0.0
pyperclip~=1.8.2
Pillow~=8.4.0